# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: JSS Exam Analyser API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List students, optionally filtered by class
      parameters:
        - name: classId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Create a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Delete a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /scores:
    get:
      operationId: listScores
      tags: [scores]
      summary: List scores for a given exam (and optionally a student)
      parameters:
        - name: examId
          in: query
          required: true
          schema:
            type: integer
        - name: studentId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Score"
    post:
      operationId: upsertScores
      tags: [scores]
      summary: Upsert scores in bulk (for a student in an exam)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScoreBulkInput"
      responses:
        "200":
          description: Upserted scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Score"
  /reports/{examId}/{studentId}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get full report for a student in an exam
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentReport"
    patch:
      operationId: updateReport
      tags: [reports]
      summary: Update comments on a student report
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportCommentInput"
      responses:
        "200":
          description: Updated report comments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentReport"
...
components:
  schemas:
...
    School:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: ["string", "null"]
        motto:
          type: ["string", "null"]
        principalName:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
...
    Student:
      type: object
      required: [id, name, admissionNo, classId]
      properties:
        id:
          type: integer
        name:
          type: string
        admissionNo:
          type: string
        classId:
          type: integer
        className:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
    StudentInput:
      type: object
      required: [name, admissionNo, classId]
      properties:
        name:
          type: string
        admissionNo:
          type: string
        classId:
          type: integer
        gender:
          type: string
        dateOfBirth:
          type: string
    StudentUpdate:
      type: object
      properties:
        name:
          type: string
        admissionNo:
          type: string
        classId:
          type: integer
        gender:
          type: string
        dateOfBirth:
          type: string
...
    Exam:
      type: object
      required: [id, name, classId, year, term]
      properties:
        id:
          type: integer
        name:
          type: string
        classId:
          type: integer
        className:
          type: ["string", "null"]
        year:
          type: integer
        term:
          type: integer
        openingDate:
          type: ["string", "null"]
        closingDate:
          type: ["string", "null"]
        status:
          type: string
...
    Score:
      type: object
      required: [id, studentId, examId, learningAreaId, marks]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        examId:
          type: integer
        learningAreaId:
          type: integer
        learningAreaName:
          type: ["string", "null"]
        learningAreaAbbr:
          type: ["string", "null"]
        marks:
          type: number
        maxMarks:
          type: integer
        rubricGrade:
          type: string
        rubricPoints:
          type: number
    ScoreEntry:
      type: object
      required: [learningAreaId, marks]
      properties:
        learningAreaId:
          type: integer
        marks:
          type: number
    ScoreBulkInput:
      type: object
      required: [studentId, examId, scores]
      properties:
        studentId:
          type: integer
        examId:
          type: integer
        scores:
          type: array
          items:
            $ref: "#/components/schemas/ScoreEntry"
    SubjectPerformance:
      type: object
      required: [learningAreaId, learningAreaName, abbreviation, marks, maxMarks, rubricGrade, rubricPoints, percentage]
      properties:
        learningAreaId:
          type: integer
        learningAreaName:
          type: string
        abbreviation:
          type: string
        marks:
          type: number
        maxMarks:
          type: integer
        rubricGrade:
          type: string
        rubricPoints:
          type: number
        percentage:
          type: number
    StudentReport:
      type: object
      required: [student, exam, school, subjects, totalMarks, totalMaxMarks, averagePercentage, averagePoints, overallGrade, rank, classSize]
      properties:
        student:
          $ref: "#/components/schemas/Student"
        exam:
          $ref: "#/components/schemas/Exam"
        school:
          $ref: "#/components/schemas/School"
        subjects:
          type: array
          items:
            $ref: "#/components/schemas/SubjectPerformance"
        totalMarks:
          type: number
        totalMaxMarks:
          type: integer
        averagePercentage:
          type: number
        averagePoints:
          type: number
        overallGrade:
          type: string
        rank:
          type: integer
        classSize:
          type: integer
        teacherComment:
          type: ["string", "null"]
        principalComment:
          type: ["string", "null"]
    ReportCommentInput:
      type: object
      properties:
        teacherComment:
          type: string
        principalComment:
          type: string
...