# lib/api-zod/src/generated/types/scoreEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JSS Exam Analyser API
 * OpenAPI spec version: 0.1.0
 */

export interface ScoreEntry {
  learningAreaId: number;
  marks: number;
}