# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JSS Exam Analyser API
 * OpenAPI spec version: 0.1.0
 */

export interface Student {
  id: number;
  name: string;
  admissionNo: string;
  classId: number;
  /** @nullable */
  className?: string | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  dateOfBirth?: string | null;
}