# lib/api-zod/src/generated/types/studentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JSS Exam Analyser API
 * OpenAPI spec version: 0.1.0
 */

export interface StudentInput {
  name: string;
  admissionNo: string;
  classId: number;
  gender?: string;
  dateOfBirth?: string;
}